fix: fix resource problem and move classes
