+++ src/main/java/indi/wenyan/setup/datagen/recipe/CheckerRecipeProvider.java
@@ -1,11 +1,10 @@
 package indi.wenyan.setup.datagen.recipe;
 
 import com.mojang.logging.annotations.MethodsReturnNonnullByDefault;
-import indi.wenyan.setup.datagen.recipe.crafting.CraftingHandrunnerRecipes;
-import indi.wenyan.setup.datagen.recipe.wyquestion.BlockRecipes;
+import indi.wenyan.setup.datagen.recipe.crafting.BlockRecipes;
 import indi.wenyan.setup.datagen.recipe.wyquestion.HandrunnerRecipes;
-import indi.wenyan.setup.datagen.recipe.wyquestion.ItemRecipes;
-import indi.wenyan.setup.datagen.recipe.wyquestion.ModuleRecipes;
+import indi.wenyan.setup.datagen.recipe.crafting.ItemRecipes;
+import indi.wenyan.setup.datagen.recipe.crafting.ModuleRecipes;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.RecipeOutput;
@@ -46,7 +45,7 @@ public class CheckerRecipeProvider extends RecipeProvider {
     @Override
     protected void buildRecipes() {
         HandrunnerRecipes.build(this.items, output, this);
-        CraftingHandrunnerRecipes.build(this.items, output, this);
+        indi.wenyan.setup.datagen.recipe.crafting.HandrunnerRecipes.build(this.items, output, this);
         ModuleRecipes.build(this.items, output, this);
         ItemRecipes.build(items, output, this);
         BlockRecipes.build(this.items, output, this);
